mostly working
